refactor(web): extract App subcomponents (CS-6 PR5)
Extract five presentational subcomponents from App.tsx into
components/app/, each a self-contained file with explicit props:

- FilterChip: reusable toggle button for filter values
- BrowseByToggle: projects/agents radio group
- SidebarFlatSessionList: bookmarked/session list with agent icons
- SearchFilterBar: the full filter scope/agent/tag/signal/cost UI
  (owns the SMART_TAG/SEARCH_TOOL/FILE_ACTIVITY/COST_RANGE constants)
- SearchResultsPanel: search results list + filter bar + loading/empty states
  (owns SEARCH_MATCH_LABELS + toSafeSnippetHtml)

Shared types (BrowseBy, CostRangeId, SearchFilterState, SearchProjectOption)
move to components/app/types.ts. App.tsx: -478 lines.

Closes CS-6 (final PR of the multi-PR sequence).
